+++ packages/compiler/src/index.ts
@@ -2,6 +2,7 @@ import compile from './compile';
 export { default as getExports } from './get-exports';
 
 export * from './util';
+export * as transformers from './transforms/index';
 export type { TransformOptions } from './transform';
 export type { Options } from './compile';
 export default compile;